Fix #3
