+++ models/admin.js
@@ -19,12 +19,12 @@ const adminSchema=new admin_Schema({
 })
 
 // adminSchema.plugin(mongooseAutoInc.plugin, 'admin');
-mongooseAutoInc.initialize(mongoose.connection);
-adminSchema.plugin(mongooseAutoInc.plugin, {
-    model: 'adminSchema',
-    field: '_id',
-    startAt: 1,
-    incrementBy: 1
-});
+// mongooseAutoInc.initialize(mongoose.connection);
+// adminSchema.plugin(mongooseAutoInc.plugin, {
+//     model: 'adminSchema',
+//     field: '_id',
+//     startAt: 1,
+//     incrementBy: 1
+// });
 
 module.exports = mongoose.model('admin',adminSchema);
