+++ models/autoReport.js
@@ -11,11 +11,11 @@ const autoReportSchema=new autoReport_Schema({
 })
 
 // autoReportSchema.plugin(mongooseAutoInc.plugin, 'autoReport');
-mongooseAutoInc.initialize(mongoose.connection);
-autoReportSchema.plugin(mongooseAutoInc.plugin, {
-    model: 'autoReportSchema',
-    field: '_id',
-    startAt: 1,
-    incrementBy: 1
-});
+// mongooseAutoInc.initialize(mongoose.connection);
+// autoReportSchema.plugin(mongooseAutoInc.plugin, {
+//     model: 'autoReportSchema',
+//     field: '_id',
+//     startAt: 1,
+//     incrementBy: 1
+// });
 module.exports = mongoose.model('autoReport',autoReportSchema);
