+++ models/review.js
@@ -24,11 +24,11 @@ const reviewSchema=new review_Schema({
 })
 
 // reviewSchema.plugin(mongooseAutoInc.plugin, 'review');
-mongooseAutoInc.initialize(mongoose.connection);
-reviewSchema.plugin(mongooseAutoInc.plugin, {
-    model: 'reviewSchema',
-    field: '_id',
-    startAt: 1,
-    incrementBy: 1
-});
+// mongooseAutoInc.initialize(mongoose.connection);
+// reviewSchema.plugin(mongooseAutoInc.plugin, {
+//     model: 'reviewSchema',
+//     field: '_id',
+//     startAt: 1,
+//     incrementBy: 1
+// });
 module.exports = mongoose.model('review',reviewSchema);
