+++ models/trade.js
@@ -18,7 +18,8 @@ const tradeSchema=new trade_Schema({
         type:String,
         required:true
     },
-    //거래장소
+    // 자동신고가 가능하게 자세히 ...
+    // 거래장소 ex) 경기도 수원시 영통구 중부대로271번길, 27-9 104-1402
     location:{
         type:String,
         required:true
@@ -38,11 +39,11 @@ const tradeSchema=new trade_Schema({
 })
 
 // tradeSchema.plugin(mongooseAutoInc.plugin, 'trade');
-mongooseAutoInc.initialize(mongoose.connection);
-tradeSchema.plugin(mongooseAutoInc.plugin, {
-    model: 'tradeSchema',
-    field: '_id',
-    startAt: 1,
-    incrementBy: 1
-});
+// mongooseAutoInc.initialize(mongoose.connection);
+// tradeSchema.plugin(mongooseAutoInc.plugin, {
+//     model: 'tradeSchema',
+//     field: '_id',
+//     startAt: 1,
+//     incrementBy: 1
+// });
 module.exports = mongoose.model('trade',tradeSchema);
